pyproject.toml
[tool.poetry]
name = "thumpy"
version = "0.1.0"
description = ""
authors = ["MooGoo <moogoo78@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.0.1"
click = "^7.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
